% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.decision.R
\name{conduct.phase3}
\alias{conduct.phase3}
\title{Function to estimate the hazard ratios and other statistics of the selected
subgroup}
\usage{
conduct.phase3(data, eta, theta)
}
\arguments{
\item{data}{a data.frame in which to interpret the variables named in the formula}

\item{eta}{a cutoff probability for the strength of evidence for decision-making}

\item{theta}{a clinically meaningful treatment effect size defined by clinicians}
}
\value{
conduct.phase3()
}
\description{
This function is used to estimate the effect size of the selected
subgroup.
}
\examples{
\donttest{conduct.phase3(example.2,eta=0.8, theta=0.95)}


}
