% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estss.R
\name{estss}
\alias{estss}
\title{Estimate the specificity and sensitivity of a somatic mutation based on its read-count data.}
\usage{
estss(A, D, O, pi, varphi)
}
\arguments{
\item{A}{A vector for the number of alternative reads for the somatic mutation.}

\item{D}{A vector for the read-depths for the somatic mutation.}

\item{O}{A vector for the observed somatic mutation values.}

\item{pi}{The mean parameters for the beta-binomial distributions when the true somatic mutation value equals 0 or 1.}

\item{varphi}{The over-dispersion parameters for the beta-binomial distributions when the true somatic mutation value equals 0 or 1.}
}
\value{
The estimated specificity and sensitivity for the somatic mutaiton.
}
\description{
Estimate the specificity and sensitivity of a somatic mutation based on its read-count data.
}
