% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcmml.R
\name{fcmml}
\alias{fcmml}
\title{Fisher score method for estimating the regression coeffcients in a mixture of logisitc regression.}
\usage{
fcmml(Y, X, alpha, beta, T0S, T1S, converged = 1e-06, maxIT = 200)
}
\arguments{
\item{Y}{The response variable. Must be binary.}

\item{X}{The design matrix. Intercept included.}

\item{alpha}{The initial values of the coefficients for the confounders.}

\item{beta}{The initial value of the coefficient for the true somatic mutation value.}

\item{T0S}{The weights vector for the mixture when the true somatic muation equals 0.}

\item{T1S}{The weights vector for the mixture when the true somatic muation equals 1.}

\item{converged}{The tolerance for the convergence. Default is 1e-6.}

\item{maxIT}{The maximal number of the EM iteration times. Default is 200.}
}
\value{
The estimated regression coefficients.
}
\description{
Fisher score method for estimating the regression coeffcients in a mixture of logisitc regression.
}
