% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gSAME.R
\name{gSAME}
\alias{gSAME}
\title{Association analysis using gene-level association analysis.}
\usage{
gSAME(Y, X, O, D, A, out_type = "C", theta_init, d0 = 20, null = FALSE,
  sig = NULL, spesen = NULL, mix_2bb = NULL, mix_4bb = NULL,
  min_altcount = 1, maxIt = 200, converged = 1e-06, reEst = 1,
  traceIt = 0, ...)
}
\arguments{
\item{Y}{The response variable. Could be continuous or binary.}

\item{X}{The design matrix. Intercept included.}

\item{O}{A matrix for the observed somatic mutation.}

\item{D}{A matrix for the total read-depth.}

\item{A}{A matrix for the number of alternative number matrix.}

\item{out_type}{The outcome type, "C" for continous, "D" for dichotomous. Default is "C".}

\item{theta_init}{The initail values of the parameters. Can be NULL.}

\item{d0}{The minimum of the total read-depth for obtaining the observed somatic mutation value. The default value is 20.}

\item{null}{Logical. Indicating the estimation using EM algorim under the null hypothesis or not. The default is FALSE.}

\item{sig}{A matrix with the likelihood of O, D, A conditioning on the true value of the somatic mutation (S=0/1) for all the samples. Default is NULL.}

\item{spesen}{A dataframe specifying the specificity and sensitivity for all the somatic mutations.}

\item{mix_2bb}{A dataframe indicating the parameters of two beta-binomial distributions depending on the true value of the somatic mutation when the read-depth is low.}

\item{mix_4bb}{A dataframe indicating the parameters of four beta-binomial distributions depending on the values of the observed somatic mutaton and the true somatic mutation when the read-depth is high.}

\item{min_altcount}{The mimimum of the number of alternative reads that the somatic mutation could acutally orrur. The default value is 1.}

\item{converged}{The tolerance for the convergence. Default is 1e-6.}

\item{mix_4bb}{A dataframe indicating the parameters of four beta-binomial distributions depending on the values of the observed somatic mutaton and the true somatic mutation when the read-depth is high.}

\item{maxIT}{The maximal number of the EM iteration times. Default is 200.}
}
\value{
A list containing the output of the EM algorithm.
\item{Theta} A matrix including the estimators of the parameters in all the iterations.
\item{theta} The estimator of the parameters.
\item{LogLik} A vector of the log-likelihood values in all the iterations.
\item{logLik} The final log-likelihood.
\item{it} The number of the iteration times for convergence
}
\description{
Association analysis using gene-level association analysis.
}
