% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSAME.R
\name{mSAME}
\alias{mSAME}
\title{Association analysis using mutation-level association analysis.}
\usage{
mSAME(Y, X, O, D, A, out_type = "C", theta_init, mix_4bb, null = FALSE,
  d0 = 20, gamma0 = 1, gamma1 = 1, bounds = NULL, maxIt = 200,
  converged = 1e-06, reEst = 1, traceIt = 0, ...)
}
\arguments{
\item{Y}{The response variable. Could be continuous or binary.}

\item{X}{The design matrix. Intercept included.}

\item{O}{A vector for the observed somatic mutation.}

\item{D}{A vector for the total read-depth.}

\item{A}{A vector for the number of alternative number matrix.}

\item{out_type}{The outcome type, "C" for continous, "D" for dichotomous. Default is "C".}

\item{theta_init}{The initail values of the parameters. Can be NULL.}

\item{mix_4bb}{A dataframe indicating the parameters of four beta-binomial distributions depending on the values of the observed somatic mutaton and the true somatic mutation when the read-depth is high.}

\item{null}{Logical. Indicating the estimation using EM algorim under the null hypothesis or not. The default is FALSE.}

\item{d0}{The minimum of the total read-depth for obtaining the observed somatic mutation value. The default value is 20.}

\item{bounds}{Some parameters for the bounds in the EM algorithm. Can be NULL.}

\item{converged}{The tolerance for the convergence. Default is 1e-6.}

\item{gamm0}{The specificity of the somatic mutation. Default is 1.}

\item{gamm1}{The sensitivity of the somatic mutation. Default is 1.}

\item{maxIT}{The maximal number of the EM iteration times. Default is 200.}
}
\value{
A list containing the output of the EM algorithm.
\item{Theta} A matrix including the estimators of the parameters in all the iterations.
\item{theta} The estimator of the parameters.
\item{LogLik} A vector of the log-likelihood values in all the iterations.
\item{logLik} The final log-likelihood.
\item{it} The number of the iteration times for convergence.
}
\description{
Association analysis using mutation-level association analysis.
}
